pyproject.toml
[tool.poetry]
name = "geekstaks"
version = "0.1.0"
description = ""
authors = ["Nurs <jackmovies01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
